- using size_t instead of int for number_of_results

+++ keyvi/include/keyvi/dictionary/completion/multiword_completion.h
@@ -49,7 +49,7 @@ public:
     }
 
     MatchIterator::MatchIteratorPair GetCompletions(
-        const std::string& query, int number_of_results = 10) const {
+        const std::string& query, size_t number_of_results = 10) const {
 
       uint64_t state = fsa_->GetStartState();
       size_t number_of_tokens;
